Refactor all of the code
