+++ src/model/sweep.py
@@ -1,4 +1,4 @@
-from src.model.ray import Ray
+from model.ray import Ray
 
 
 class Sweep:
